+++ tasks/omp/yasakova_t_sparse_matrix_multiplication/include/ops_omp.hpp
